+++ src/lib/components/data-table/DataTableEvent.svelte
@@ -56,7 +56,7 @@
 		});
 	}
 
-	let showOnlySuccess = writable(false);
+	let showRejected = writable(false);
 
 	// Reactive statement to filter payments based on search term and switch state
 	$: filteredPayments = Payments.filter((payment) => {
@@ -67,7 +67,11 @@
 			payment.payment_status?.toLowerCase().includes(searchTerm.toLowerCase()) ||
 			payment.rut?.toLowerCase().includes(searchTerm.toLowerCase());
 
-		const matchesSuccessFilter = $showOnlySuccess || payment.payment_status === 'success';
+		const matchesSuccessFilter = $showRejected || payment.payment_status === 'success';
+
+		if ($showRejected) {
+			return payment.payment_status !== 'success';
+		}
 
 		return matchesSearchTerm && matchesSuccessFilter;
 	});
@@ -77,10 +81,10 @@
 <div class="flex flex-col md:flex-row items-center py-4 md:gap-4 justify-between">
 	<div class="flex flex-col md:flex-row md:items-center gap-4 items-left w-full">
 		<h2 class="font-bold text-xl">Listado de compras</h2>
-		<Input type="text" placeholder="Buscador..." class="w-full md:w-96" bind:value={searchTerm} />
+		<Input type="text" placeholder="Buscador..." class="w-full md:w-96 border-primary" bind:value={searchTerm} />
 		<div class="flex items-center space-x-2">
-			<Switch id="rejected-payments" bind:checked={$showOnlySuccess} />
-			<Label for="rejected-payments">Incluir pagos rechazados</Label>
+			<Switch id="rejected-payments" bind:checked={$showRejected} />
+			<Label for="rejected-payments">Ver pagos rechazados</Label>
 		</div>
 	</div>
 	<DialogToAddPayments />